requirements.txt
openai>=1.0.0
requests>=2.31.0
edge-tts>=6.1.9
Pillow>=10.0.0
moviepy>=2.0.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
tavily-python>=0.3.0
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.12.0
